Rename .js files to lowercase, words are separated by hyphens
